refactor: heroContents를 HeroSection으로 이동
- Carousel prop에도 heroContents를 사용할 수 있게 함
